+++ src/core/client/systems/index.ts
@@ -6,6 +6,7 @@ export * as interaction from './interaction';
 export * as inventory from './inventory';
 export * as messenger from './messenger';
 export * as playerConfig from './playerConfig';
+export * as plugins from './plugins';
 export * as rpc from './rpc';
 export * as sound from './sound';
 export * as wheelMenu from '../views/wheelMenu';